+++ components/listings/show/Head.js
@@ -24,6 +24,7 @@ export default class ListingHead extends Component {
         imageSrc={seoImgSrc}
         imageWidth={'1024'}
         imageHeight={'768'}
+        url={'https://www.emcasa.com/'}
       >
         <link
           rel="stylesheet"
